refactor QueryMetadata and PersistentQueryMetadata
